Аdded photo rating function and some other functionality was changed

+++ src/entity/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Date, Integer, ForeignKey, DateTime, func, Boolean, Column, Table
+from sqlalchemy import String, Date, Integer, ForeignKey, DateTime, func, Boolean, Column, Table, Float
 from sqlalchemy.orm import Mapped, mapped_column, DeclarativeBase, relationship
 from pydantic import EmailStr
 from datetime import date
@@ -36,7 +36,7 @@ class Photo(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     image: Mapped[str] = mapped_column(String(255), nullable=False)
     description: Mapped[Optional[str]] = mapped_column(String(250))
-    rating: Mapped[int] = mapped_column(Integer, nullable=True)
+    rating: Mapped[float] = mapped_column(Float, nullable=True)
     user_id: Mapped[int] = mapped_column(Integer, ForeignKey('user.id'), nullable=False)
     user: Mapped["User"] = relationship("User", backref="photo", lazy="joined")
     created_at: Mapped[date] = mapped_column('created_at', DateTime, default=func.now(), nullable=True)
